requirements.txt
numpy
torch
pyyaml
opacus

# Optional (only needed for privacy curve plots)
# matplotlib

# Optional (needed for pca_dim/standardize and RandomForest baselines)
# scikit-learn

# Optional (only needed for CAMELYON17-WILDS prep script)
# wilds
# torchvision

# Optional (only needed for BrainSCOPE prep script)
# pandas
